Fix accuracy measurement

+++ gram_schmidt.c
@@ -4,18 +4,18 @@
 #include <math.h>
 
 // Function to get Gram Schmidt information from set of basis vectors
-gs_info gram_schmidt(double ** basis, int N){
+gs_info gram_schmidt(long double ** basis, int N){
 
     // initialise new basis and allocate necessary memory
-    double ** gs_basis = malloc(sizeof(double*)*N);
+    long double ** gs_basis = malloc(sizeof(long double*)*N);
 
     // initialise the gram schmidt coefficients and allocate necessary memory
-    double ** mu = malloc(sizeof(double)*N);
+    long double ** mu = malloc(sizeof(long double)*N);
     
     // define mu and gs_basis to have N arrays of size N and allocate necessary memory
     for (int i = 0; i < N; i++){
-        mu[i] = malloc(sizeof(double)*N);
-        gs_basis[i] = malloc(sizeof(double)*N);
+        mu[i] = malloc(sizeof(long double)*N);
+        gs_basis[i] = malloc(sizeof(long double)*N);
     }
 
     // main loop for gram schmidt
